fix: Refactor bff context and hooks into react-spa shared
