+++ libs/react-spa/bff/src/index.ts
@@ -1,5 +1,8 @@
-export * from './lib/BffContext'
 export * from './lib/BffProvider'
-export * from './lib/bff.hooks'
 export * from './lib/bff.mocks'
 export * from './lib/bff.utils'
+export {
+  useUserInfo,
+  useAuth,
+  useBffUrlGenerator,
+} from '@island.is/react-spa/shared'
