+++ libs/react-spa/bff/src/lib/bff.state.ts
@@ -1,14 +1,10 @@
+import {
+  BffError,
+  NonLoggedInState,
+  BffReducerState,
+  NonLoggedInAuthState,
+} from '@island.is/react-spa/shared'
 import { BffUser } from '@island.is/shared/types'
-import { BffError } from './BffError'
-
-// Defining the possible states for authentication
-export type BffState =
-  | 'logged-out'
-  | 'loading'
-  | 'logged-in'
-  | 'switching'
-  | 'logging-out'
-  | 'error'
 
 export enum ActionType {
   SIGNIN_START = 'SIGNIN_START',
@@ -18,29 +14,6 @@ export enum ActionType {
   SWITCH_USER = 'SWITCH_USER',
   ERROR = 'ERROR',
 }
-
-type NonLoggedInAuthState = Exclude<BffState, 'logged-in'>
-
-export interface BffReducerStateBase {
-  authState: BffState
-  isAuthenticated: boolean
-  error?: BffError | null
-}
-
-// State when the user is not logged in
-export interface NonLoggedInState extends BffReducerStateBase {
-  authState: NonLoggedInAuthState
-  userInfo: null
-}
-
-// State when the user is logged in
-export interface LoggedInState extends BffReducerStateBase {
-  authState: 'logged-in'
-  userInfo: BffUser
-  isAuthenticated: true
-}
-
-export type BffReducerState = NonLoggedInState | LoggedInState
 
 export const initialState: NonLoggedInState = {
   userInfo: null,
