+++ libs/react-spa/shared/src/bff/BffError.ts
