+++ libs/react-spa/shared/src/index.ts
@@ -16,3 +16,9 @@ export * from './hooks/useBroadcaster'
 // utils
 export * from './utils/getOrganizationSlugFromError'
 export * from './utils/notFoundError'
+
+// bff
+export * from './bff/BffContext'
+export * from './bff/bff.hooks'
+export * from './bff/bff.state'
+export * from './bff/BffError'
